backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==12.1.1
pyotp==2.9.0
email-validator==2.0.0
marshmallow==3.20.1
Werkzeug==2.3.7
sendgrid==6.10.0
